requirements.txt
PyGithub==2.1.1
python-dotenv==1.0.0
pytest==7.4.3
pydantic==1.10.13
aiohttp==3.11.11
asyncio==3.4.3
PyYAML==6.0.1
pytest-asyncio==0.23.5
